[NFC] Split ModuleTrace Infrastructure out of FrontendTool
